[REF] get_time_spent in analytic.account
