+++ contract_timesheet_monitoring/models/__init__.py
@@ -1,2 +1,3 @@
 from . import contract_line
 from . import account_move_line
+from . import account_analytic_account
